# Speed up CLI loading by reducing the number of explicitly loaded files