--- a/cli/lib/kontena/callbacks/auth/01_list_and_select_grid_after_master_auth.rb
+++ b/cli/lib/kontena/callbacks/auth/01_list_and_select_grid_after_master_auth.rb
@@ -2,8 +2,6 @@ module Kontena
   module Callbacks
     class ListAndSelectGrid < Kontena::Callback
 
-      include Kontena::Cli::Common
-
       matches_commands 'master login'
 
       def after_load
@@ -15,6 +13,7 @@ module Kontena
       # Runs kontena grids list --use which will auto join the first available
       # grid
       def after
+        extend Kontena::Cli::Common
         return if command.skip_grid_auto_select?
         return unless current_master
         return unless command.exit_code == 0