--- a/cli/lib/kontena/callbacks/master/deploy/40_install_ssl_certificate_after_deploy.rb
+++ b/cli/lib/kontena/callbacks/master/deploy/40_install_ssl_certificate_after_deploy.rb
@@ -2,11 +2,11 @@ module Kontena
   module Callbacks
     class InstallSslCertificateAfterDeploy < Kontena::Callback
 
-      include Kontena::Cli::Common
-
       matches_commands 'master create'
 
       def after
+        extend Kontena::Cli::Common
+
         return unless command.exit_code == 0
         return unless command.result.kind_of?(Hash)
         return unless command.result.has_key?(:ssl_certificate)