--- a/cli/lib/kontena/callbacks/master/deploy/50_authenticate_after_deploy.rb
+++ b/cli/lib/kontena/callbacks/master/deploy/50_authenticate_after_deploy.rb
@@ -5,6 +5,8 @@ module Kontena
       matches_commands 'master create'
 
       def after
+        extend Kontena::Cli::Common
+
         require 'securerandom'
         extend Kontena::Cli::Common
         logger.debug { "Command result: #{command.result.inspect}" }