--- a/cli/lib/kontena/callbacks/master/deploy/55_create_initial_grid_after_deploy.rb
+++ b/cli/lib/kontena/callbacks/master/deploy/55_create_initial_grid_after_deploy.rb
@@ -2,11 +2,11 @@ module Kontena
   module Callbacks
     class CreateInitialGridAfterDeploy < Kontena::Callback
 
-      include Kontena::Cli::Common
-
       matches_commands 'master create'
 
       def after
+        extend Kontena::Cli::Common
+
         return unless command.exit_code == 0
         return unless config.current_master
         return unless config.current_master.name == command.result[:name]