--- a/cli/lib/kontena/callbacks/master/deploy/60_configure_auth_provider_after_deploy.rb
+++ b/cli/lib/kontena/callbacks/master/deploy/60_configure_auth_provider_after_deploy.rb
@@ -2,8 +2,6 @@ module Kontena
   module Callbacks
     class ConfigureAuthProviderAfterDeploy < Kontena::Callback
 
-      include Kontena::Cli::Common
-
       matches_commands 'master create'
 
       def init_cloud_args
@@ -16,6 +14,7 @@ module Kontena
       end
 
       def after
+        extend Kontena::Cli::Common
         return unless command.exit_code == 0
         return unless command.result.kind_of?(Hash)
         return unless command.result.has_key?(:name)