--- a/cli/lib/kontena/cli/apps/kontena_yml_generator.rb
+++ b/cli/lib/kontena/cli/apps/kontena_yml_generator.rb
@@ -1,5 +1,6 @@
-require 'yaml'
 require_relative 'common'
+require "safe_yaml"
+SafeYAML::OPTIONS[:default_mode] = :safe
 
 module Kontena::Cli::Apps
   class KontenaYmlGenerator