--- a/cli/lib/kontena/cli/apps/yaml/reader.rb
+++ b/cli/lib/kontena/cli/apps/yaml/reader.rb
@@ -1,4 +1,5 @@
-require 'yaml'
+require "safe_yaml"
+SafeYAML::OPTIONS[:default_mode] = :safe
 require_relative 'service_extender'
 require_relative 'validator'
 require_relative 'validator_v2'