--- a/cli/lib/kontena/cli/apps/yaml/service_extender.rb
+++ b/cli/lib/kontena/cli/apps/yaml/service_extender.rb
@@ -1,4 +1,3 @@
-require 'yaml'
 require_relative '../../../util'
 
 module Kontena::Cli::Apps