--- a/cli/lib/kontena/cli/stacks/registry/show_command.rb
+++ b/cli/lib/kontena/cli/stacks/registry/show_command.rb
@@ -14,6 +14,8 @@ module Kontena::Cli::Stacks::Registry
 
     def execute
       require 'semantic'
+      require "safe_yaml"
+      SafeYAML::OPTIONS[:default_mode] = :safe
       unless versions?
         stack = ::YAML.safe_load(stacks_client.show(stack_name, stack_version))
         puts "#{stack['stack']}:"