--- a/cli/lib/kontena/cli/vault/export_command.rb
+++ b/cli/lib/kontena/cli/vault/export_command.rb
@@ -11,6 +11,9 @@ module Kontena::Cli::Vault
 
     def execute
       require 'shellwords'
+      require 'json'
+      require "safe_yaml"
+      SafeYAML::OPTIONS[:default_mode] = :safe
       meth = json? ? :to_json : :to_yaml
       puts(
         Kontena.run!(['vault', 'ls', '--return']).sort.map do |secret|