--- a/cli/lib/kontena/cli/vault/import_command.rb
+++ b/cli/lib/kontena/cli/vault/import_command.rb
@@ -15,6 +15,9 @@ module Kontena::Cli::Vault
     requires_current_master
 
     def parsed_input
+      require "json"
+      require "safe_yaml"
+      SafeYAML::OPTIONS[:default_mode] = :safe
       json? ? JSON.load(input) : YAML.safe_load(input)
     end