Before: (with #2334 and the plugin disabler from this PR)

```
$ time NO_PLUGINS=true bin/kontena whoami --help
0.26s user 0.06s system 92% cpu 0.342 total
```

After:
```
$ time NO_PLUGINS=true bin/kontena whoami --help
0.09s user 0.04s system 84% cpu 0.161 total
```