@jnummelin 2017-05-26
Needs rebase/merge