@jnummelin 2017-05-26
> Right. It now needs #2263

So this depends now on #2263 to be merged in first?