@kke 2017-05-26
Snappy!

![image](https://cloud.githubusercontent.com/assets/224971/26496230/541d8126-422f-11e7-93c4-c7c6f2682f1e.png)

However, plugins are slooooow to load:
![image](https://cloud.githubusercontent.com/assets/224971/26496259/70ab5d36-422f-11e7-8e65-1285453c7618.png)